Parts Editor: make the connectors layer, the main one (not a layerkin)
git-svn-id: http://fritzing.googlecode.com/svn/trunk@2412 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/partseditor/partseditormainwindow.cpp
@@ -565,10 +565,10 @@ void PartsEditorMainWindow::cleanUp() {
 			m_closedBeforeCount++;
 		}
 	}
-	/*if(m_tempDir.path() != ".") {
+	if(m_tempDir.path() != ".") {
 		rmdir(m_tempDir);
 		m_tempDir = QDir();
-	}*/
+	}
 }
 
 void PartsEditorMainWindow::parentAboutToClose() {
